Split up lib/fake_require.js
There is no shared client/server code in the file, so why not make two files?
